PR Update: Refactor to build in a release pipeline
Coherency can be an issue when attempting to base a release on
artifacts from CI pipelines.

This change moves the release to a dedicated pipeline that performs
builds on the code that triggers the release and only downloads
files from that release as well.

Some names and other modifications are made given this approach.